Acrescimo da classe Participante.
Obs: para melhorar a criação dos metodos, pois as ias e o jogador são todos participantes da partida.
